인증객체테스트완
