+++ src/main/java/com/jjjteam/jmarket/controller/HomeController.java
@@ -1,9 +1,13 @@
 package com.jjjteam.jmarket.controller;
 
+import com.jjjteam.jmarket.security.services.UserDetailsImpl;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
+@Slf4j
 public class HomeController {
 
 	// 사이트?
@@ -31,14 +35,7 @@ public class HomeController {
 	}
 
 	@GetMapping("/index")
-	public String toLoginPage2() throws Exception {
-
-		try {
-
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
-		}
-
+	public String toLoginPage2()  {
 		return "index";
 	}
